WeaponSwithing on scrollbar, also removes Scope. Reload < maxAmmo

+++ Assets/Dungeons & Droids/Scripts/PlayerManager.cs
@@ -68,7 +68,7 @@ public class PlayerManager : MonoBehaviour
         }
     }
 
-    public IEnumerator SelectedWeapon(float weaponIndex)
+    public IEnumerator SelectedWeapon(int weaponIndex)
     {
         if (weaponIndex == 1)
         {
